// client/src/types/aseprite.ts
export interface Dimensions {
  w: number
  h: number
}

export interface Coordinates {
  x: number
  y: number
}

export interface Box extends Dimensions, Coordinates {}

export interface Frame {
  frame: Box
  rotated: boolean
  trimmed: boolean
  spriteSourceSize: Box
  sourceSize: Dimensions
  duration: number
}

export interface Slice {
  /**
   * 'south' | 'southwest' | 'west' | 'northwest' | 'north'
   */
  name: string
  color: string
  keys: {
    frame: number
    bounds: Box
  }[]
}

/**
 * The top-level output format generated by the aseprite command.
 */
export interface AsepriteFile {
  frames: {
    /**
     * Key is in the format: `{tag}:{frame}:{layer}`, and by convention,
     * the layers are named `{type}.{name}[.{optionalSubName}]
     */
    [key: string]: Frame
  }
  meta: {
    size: Dimensions
    frameTags: {
      name: string
      from: number
      to: number
      direction: 'forward' | 'pingpong'
    }[]
    slices: Slice[]
  }
}